requirements.txt
requests
python-dotenv
sounddevice
wavio
google-genai
pvporcupine